Information.md
Take into account that it is not the final version of the engine, this game engine has a good potential, so I will continue to develop it little by little with new updates. For now, I can only recommend you try the tests of "Mario" which is a mini platform with some bugs that I will correct and "Snake" which is the classic snake game only with a simple reappearance bug that I will fix soon, also I plan to add a "mini programming language" which will be used to program the games much easier I also have to implement a graphical interface. Without further ado, I hope you enjoy programming with my game engine created with gpt chat and (Some of my corrections).