Update README.md
*Version 0.0.0.2 Python Implementation is available
Hello, this repository is for my game engine. This game engine is being built with chatGPT. It is a test and I hope you send me your projects with my engine, I will actively work to improve it. Suggestions are accepted.

To use my game engine, you need to recommend these things:
* Have Visual Studio code installed
* A medium to advanced knowledge of python and its pygame and pygame_gui modules.

Please note that it is not the final version of the engine, this game engine has good potential, so I will continue to develop it little by little with new updates. For now I can only recommend that you try it, in the github directory there is a .bat file that has the instructions to be pasted into the windows console.